refactor: extract pump/body/system mixins from ICModelController (step 3)
Move pump, body-temperature, and system (vacation + object getters) helpers into src/pyintellicenter/_mixins/, composed into ICModelController. Behavior-preserving: public API and __all__ unchanged, signatures identical. Heater methods and the cross-cutting get_all_entities/get_featured_entities aggregators remain in controller.py.

Also restore backward-compatible re-exports of attribute/type constants from the controller module namespace (preserving `pyintellicenter.controller.<NAME>` access) and add tests/test_controller_namespace_compat.py to guard it.
